Include install.md in index.md documentation home page
Now that the process to install Codacy on Kubernetes is simpler, we can
include the instructions on installing Codacy directly on the home page.
This makes for a bit longer home page, but simplifies the structure of
the documentation and makes it easier for users to follow the
instructions because there are less jumps between pages.