Remove links from index.md intro
I believe these links can cause extra confusion because they make the
user jump into the the sections of the same page but there are also
separate pages for some of the main steps of the installation process.
It's easy to lose track of where the user is, and the less links and
jumps there are, the better.